Move constant's definition code
